Refactor user-related classes and introduce UserController and UserService for user management
