+++ FeedBloom/src/main/java/sams/feedbloom/user/entity/User.java
@@ -1,4 +1,4 @@
-package sams.feedbloom.authentication.entity;
+package sams.feedbloom.user.entity;
 
 import jakarta.persistence.*;
 import lombok.Data;
