+++ FeedBloom/src/main/java/sams/feedbloom/user/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package sams.feedbloom.authentication.repository;
+package sams.feedbloom.user.repository;
 
 import org.springframework.stereotype.Repository;
-import sams.feedbloom.authentication.entity.User;
 import sams.feedbloom.common.repository.CommonRepository;
+import sams.feedbloom.user.entity.User;
 
 import java.util.Optional;
